+++ sunpy/net/SPICE/PSP/__init__.py
@@ -1,2 +1,1 @@
-from sunpy.net.SPICE.PSP.attrs import *
 from sunpy.net.SPICE.PSP.psp import *
