+++ sunpy/net/SPICE/Solo/__init__.py
@@ -1,2 +1,1 @@
-from sunpy.net.SPICE.Solo.attrs import *
 from sunpy.net.SPICE.Solo.solar_orbiter import *
